renamed components, changed product to products in combine reducer, get request for products 500 error
