+++ client/components/SingleProduct.js
@@ -50,7 +50,7 @@ const SingleProduct = (props)=>{
 
 const MapState = (state, ownProps)=>{
   return {
-    products: state.product,
+    products: state.products,
     productId: ownProps.match.params.productId
   }
 }
