+++ client/store/index.js
@@ -3,11 +3,11 @@ import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import user from './user'
-import product from './product'
+import products from './products'
 import cart from './cart'
 import reviews from './reviews'
 
-const reducer = combineReducers({user, cart, product, reviews})
+const reducer = combineReducers({user, cart, products, reviews})
 
 const middleware = composeWithDevTools(applyMiddleware(
   thunkMiddleware,
@@ -17,6 +17,6 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
-export * from './product'
+export * from './products'
 export * from './cart'
 export * from './reviews'
